requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
drf-spectacular==0.22.1
django-filter==22.1
python-dotenv==0.21.0
gunicorn==20.1.0
whitenoise==6.2.0
pytest
